+++ backend/src/routes/user.routes.ts
@@ -1,23 +1,11 @@
 import express from 'express';
-import {
-	createNewUser,
-	deleteUserByID,
-	getUserByID,
-	loginUser,
-} from '../controllers/user.controller';
-import { deleteRefreshTokens, useRefreshToken } from '../services/user.services';
+import { createNewUser, deleteUserByID, getUserByID } from '../controllers/user.controller';
 
 const router = express.Router();
 
 router.get('/:id', getUserByID);
 
 router.put('/create', createNewUser);
-
-router.post('/login', loginUser);
-
-router.post('/token', useRefreshToken);
-
-router.delete('/logout', deleteRefreshTokens);
 
 router.delete('/:id', deleteUserByID);
